requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
psycopg2-binary==2.9.9
python-telegram-bot==21.1.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1